Squashed commit of the following:
    -Update call to IRecoveryStrategy#recover to new signature. Update recovery strategies to use new JSinsertionRecoveryStrategy where possible.
    -Expand the IRecoveryStrategy#recover method to include the lastToken since the parser is the only place we can get that and now our recovery strategies can live outside the parser.
    -Add a unit test for the case that fails when we have an unclosed function call
    -A convenience class for JS recovery strategies which are based on simple insertion. This essentially defines the generic ITypePredicate type used by JS
    -Create a base parser error recovery strategy. This covers almost every recovery strategy being used in JS, so it's been genericized with the hope that it will be useful for other languages as well.
    -Extend ITypePredicate interface to include a method for grabbing an enumeration's Beaver id.

+++ plugins/com.aptana.editor.html/src/com/aptana/editor/html/parsing/lexer/HTMLTokenType.java
@@ -35,7 +35,7 @@ public enum HTMLTokenType implements ITypePredicate
 
 	private static final Map<String, HTMLTokenType> NAME_MAP;
 	private String _scope;
-	private int _index;
+	private short _index;
 
 	/**
 	 * static
@@ -43,7 +43,7 @@ public enum HTMLTokenType implements ITypePredicate
 	static
 	{
 		NAME_MAP = new HashMap<String, HTMLTokenType>();
-		int index = 0;
+		short index = 0;
 
 		for (HTMLTokenType type : EnumSet.allOf(HTMLTokenType.class))
 		{
@@ -73,12 +73,11 @@ public enum HTMLTokenType implements ITypePredicate
 		this._scope = scope;
 	}
 
-	/**
-	 * getIndex
-	 * 
-	 * @return
+	/*
+	 * (non-Javadoc)
+	 * @see com.aptana.parsing.lexer.ITypePredicate#getIndex()
 	 */
-	public int getIndex()
+	public short getIndex()
 	{
 		return this._index;
 	}
